:beetle: override (de)selection commands instead of the core method
Despite the docs stating that "It is currently possible to redefine existing commands by simply creating a command with the same name in your own plugin. This is discouraged.", this failed for the 'ToggleSelection' Core command.
Instead we need to override all the (de)selection commands using the 'on_command(...)' API

Fixes: #17